Alter schema to use integer in place of decimal for pricing

+++ schemas/postgres/schema/carts.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS carts (
   uuid        UUID NOT NULL,
   sku         VARCHAR(64) NOT NULL,
   qty         SMALLINT NOT NULL CHECK (qty >= 1 AND qty < 10000),
-  unit_price  NUMERIC(8, 4) NOT NULL CHECK (unit_price >= 0.0000),
+  unit_price  INTEGER NOT NULL CHECK (unit_price >= 0),
   created     TIMESTAMP NOT NULL DEFAULT NOW(),
   modified    TIMESTAMP NOT NULL DEFAULT NOW(),
   UNIQUE (uuid, sku),
